immediately adapt treetransformers
