+++ src/tensors/indexmanipulations.jl
@@ -641,8 +641,8 @@ function add_transform_kernel!(
                     ptriv, false, One(), Zero(), backend, allocator
                 )
             end
-            U′ = adapt_transformer(U, storagetype(tdst))
-            mul!(buffer_dst, buffer_src, transpose(StridedView(U′)))
+            U′ = Adapt.adapt(StridedView(U), storagetype(tdst))
+            mul!(buffer_dst, buffer_src, transpose(U′))
             @inbounds for (i, (f₃, f₄)) in enumerate(fusiontrees(dst))
                 TO.tensoradd!(
                     tdst[f₃, f₄], sreshape(buffer_dst[:, i], sz_src),
@@ -724,8 +724,8 @@ function add_transform_kernel!(
 
             # 2. Recoupling: buffer_dst = buffer_src * U^T  (each output tree is a linear
             #    combination of input trees weighted by the recoupling coefficients).
-            U′ = adapt_transformer(U, typeof(data_dst))
-            mul!(buffer_dst, buffer_src, transpose(StridedView(U′)))
+            U′ = Adapt.adapt(StridedView(U), typeof(data_dst))
+            mul!(buffer_dst, buffer_src, transpose(U′))
 
             # 3. Insert: scatter column i of buffer_dst into the destination, applying the
             #    actual index permutation p in the same tensoradd! call.
@@ -741,12 +741,3 @@ function add_transform_kernel!(
     TO.allocator_reset!(allocator, cp)
     return nothing
 end
-
-"""
-    adapt_transformer(U::AbstractMatrix, ::Type{A})
-
-Return a version of the basis transformation `U` that is compatible for storage type `A`.
-Default is a no-op.
-Backends (e.g. CUDA, AMDGPU) should overload this for their vector types to ensure the recoupling matrix `U` is on the correct device.
-"""
-adapt_transformer(U::AbstractMatrix, ::Type{A}) where {A} = U
